fix: resolve cleanup issues

+++ init.sh
@@ -124,19 +124,23 @@ else
 fi
 
 # Configure network for opensearch containers.
-echo "$(info) Creating new network called ${NET_NAME}..."
-echo
-command "$CONTAINER_RUNTIME" network rm -f "${NET_NAME}"
-command "$CONTAINER_RUNTIME" network create "${NET_NAME}"
-
-if [ $? -ne 0 ]; then
-  echo "$(error) Error occurred during 'network create' for ${NET_NAME}..."
+if command "$CONTAINER_RUNTIME" network inspect "${NET_NAME}" >/dev/null 2>&1; then
+  echo "$(info) Network ${NET_NAME} already exists, reusing..."
+else
+  echo "$(info) Creating new network called ${NET_NAME}..."
   echo
-  exit;
-fi
+  command "$CONTAINER_RUNTIME" network rm -f "${NET_NAME}"
+  command "$CONTAINER_RUNTIME" network create "${NET_NAME}"
 
-echo
-echo "$(info) Network ${NET_NAME} created..."
+  if [ $? -ne 0 ]; then
+    echo "$(error) Error occurred during 'network create' for ${NET_NAME}..."
+    echo
+    exit;
+  fi
+
+  echo
+  echo "$(info) Network ${NET_NAME} created..."
+fi
 
 install_in_container "${OS_APP}"
 
@@ -186,37 +190,36 @@ echo "$(info) =                                                                 
 echo "$(info) =  Install complete, get ready to rock OpenSearch!                                                                  ="
 echo "$(info) =                                                                                                                   ="
 echo "$(info) =  To send data using Fluent Bit, use the following command:                                                        ="
-echo "$(info) =    $ "$CONTAINER_RUNTIME" run --rm -it --network "${NET_NAME}" \                                                           ="
-echo "$(info) =      -v $(pwd)/fluent-bit.yaml:/fluent-bit/etc/fluent-bit.yaml \                                                  ="
+echo "$(info) =    $ "$CONTAINER_RUNTIME" run --rm -it --network "${NET_NAME}" \                                                                       ="
+echo "$(info) =      -v \$PWD/fluent-bit.yaml:/fluent-bit/etc/fluent-bit.yaml \                                                    ="
 echo "$(info) =      -c /fluent-bit/etc/fluent-bit.yaml                                                                           ="
 echo "$(info) =                                                                                                                   ="
 echo "$(info) =  Attach to the running container images with the following:                                                       ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =    $ $CONTAINER_RUNTIME attach ${OS_APP}                                                                          ="
-echo "$(info) =    $ $CONTAINER_RUNTIME attach ${OSD_APP}                                                                         ="
+echo "$(info) =    $ $CONTAINER_RUNTIME attach ${OS_APP}                                                                                     ="
+echo "$(info) =    $ $CONTAINER_RUNTIME attach ${OSD_APP}                                                                          ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =  The ${OS_APP} is available at:                                                                                   ="
+echo "$(info) =  The ${OS_APP} is available at:                                                                                  ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =    http://localhost:9200  (admin:${OS_PWD})                                                                       ="
+echo "$(info) =    http://localhost:9200  (admin:${OS_PWD})                                                                ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =  The ${OSD_APP} is available at:                                                                                  ="
+echo "$(info) =  The ${OSD_APP} is available at:                                                                       ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =    http://localhost:5601  (admin:${OS_PWD})                                                                       ="
+echo "$(info) =    http://localhost:5601  (admin:${OS_PWD})                                                                ="
 echo "$(info) =                                                                                                                   ="
 echo "$(info) =  If you stop the containers, they will be removed, so to restart run the following commands:                      ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =   $ $CONTAINER_RUNTIME run --name ${OS_APP} -d --network "${NET_NAME}" \                                          ="
+echo "$(info) =   $ $CONTAINER_RUNTIME run --name ${OS_APP} -d --network "${NET_NAME}" \                                                            ="
 echo "$(info) =       -p 9200:9200 -p 9600:9600               \                                                                   ="
 echo "$(info) =       -e 'discovery.type=single-node' \                                                                           ="
 echo "$(info) =       -e 'DISABLE_INSTALL_DEMO_CONFIG=true' \                                                                     ="
 echo "$(info) =       -e 'DISABLE_SECURITY_PLUGIN=true'  \                                                                        ="
-echo "$(info) =       ${OS_IMAGE}:${OS_VERSION}                                                                                   ="
+echo "$(info) =       ${OS_IMAGE}:${OS_VERSION}                                                                          ="
 echo "$(info) =                                                                                                                   ="
-echo "$(info) =   $ $CONTAINER_RUNTIME run --name ${OSD_APP} -d --network "${NET_NAME}" \                                         ="
+echo "$(info) =   $ $CONTAINER_RUNTIME run --name ${OSD_APP} -d --network "${NET_NAME}" \                                                 ="
 echo "$(info) =       -p 5601:5601 -v  -e 'DISABLE_SECURITY_DASHBOARDS_PLUGIN=true'    \                                          ="
-echo "$(info) =       -v ./${OSD_CONFIG}:/usr/share/${OSD_APP}/config/opensearch_dashboards.yml \                                 ="
-echo "$(info) =       ${OSD_IMAGE}:${OSD_VERSION}                                                                                 ="
+echo "$(info) =       -v ./${OSD_CONFIG}:/usr/share/${OSD_APP}/config/opensearch_dashboards.yml \  ="
+echo "$(info) =       ${OSD_IMAGE}:${OSD_VERSION}                                                               ="
 echo "$(info) =                                                                                                                   ="
 echo "$(info) ====================================================================================================================="
 echo
-
